+++ Model/NSEOptionChainData.cs
@@ -13,8 +13,8 @@ namespace FetchOptionChain
         public string expiryDate { get; set; }
         public string underlying { get; set; }
         public string identifier { get; set; }
-        public int openInterest { get; set; }
-        public int changeinOpenInterest { get; set; }
+        public double openInterest { get; set; }
+        public double changeinOpenInterest { get; set; }
         public double pchangeinOpenInterest { get; set; }
         public int totalTradedVolume { get; set; }
         public double impliedVolatility { get; set; }
@@ -38,8 +38,8 @@ namespace FetchOptionChain
         public string expiryDate { get; set; }
         public string underlying { get; set; }
         public string identifier { get; set; }
-        public int openInterest { get; set; }
-        public int changeinOpenInterest { get; set; }
+        public double openInterest { get; set; }
+        public double changeinOpenInterest { get; set; }
         public double pchangeinOpenInterest { get; set; }
         public int totalTradedVolume { get; set; }
         public double impliedVolatility { get; set; }
